Go back to more fragile current-user-detection
Unfortunately `ownwer_id` is not the logged in user, but the owner
of the current leaderboard. I found no other reliable way to deduce
the currently logged in user. So we resort to this hackish way after
all.

This will fail for edge cases, most obviously the one where two members
of a board have the same username. But it seems worth it?