Remove jsdoc entries (see rest of message)
Although I think these kinds of docs (or better yet: TypeScript) are
absolutely fantastic and almost necessary, they also impose a burden
on maintenance: I would need to keep them up to date, because incorrect
docs are worse than no docs.

Knowing myself and the style of this (pet) project, I don't think I am
willing to take that responsibility at the moment. So I will strip these
changes from the PR for now.

If this is a must-have for others to contribute more I am open to having
a bit further discussion in an open issue. But I'd rather not have types
for a handful of things in this project only relevant to one PR.

Hope that makes sense.

PS. I've purposely first merged these docs in, and then removed them, so
that we may grab them from history if we change our minds on this matter.